Modified the handling of the data field in the database.
There is now provision for an extended data field (edata)
which allows the data field to grow  beyond that supported
within the table key.

This version fixes this at 128 bytes except in the admin
code which detects the column width for the data column - TODO.

The table key is still limited to just the data field so
this version will fail on data with identical prefixes but
which vary in the extended data field only. This should be
fixed by adding an MD5 hash of the data as a hidden column
within the key - TODO.

Fixed up inter-process communication a bit more.

Abstracted the memory management out into a separate file!
This is in preparation for:
	o	Shared memory caches
	o	Task based storage allocation
	o	Complete dynamic storage allocation.

Some minor code clean up and fixes

Reformatting of code I have touched.

MAXDATALENGTH is set at 65536
MAXTXTLENGTH  is set at 2048 - Arbitrary but near bind
MAXTXTELEMLEN is fixed at 255 as per standards




git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@13 fff143f6-a9c2-4b89-8f93-78cb2c0259b0